+++ doc/source/changelog.rst
@@ -19,6 +19,7 @@ master: 2.5.0
 * Support non-POD by-value returns in Numba traces
 * Nullify derived class Python proxy when the C++ object is deleted
 * Add ``__cpp_template__`` back reference for instantiated templated classes
+* Improved buffer checking for ``std::initializer_list``
 
 
 2022-10-03: 2.4.1