+++ doc/source/changelog.rst
@@ -21,6 +21,7 @@ master: 2.5.0
 * Add ``__cpp_template__`` back reference for instantiated templated classes
 * Improved buffer checking for ``std::initializer_list``
 * Add convenience functions ``argc()`` and ``argv()`` to ``cppyy.ll``
+* Added ``nullptr`` comparisons for for typed ``nullptr``
 
 
 2022-10-03: 2.4.1