+++ doc/source/changelog.rst
@@ -22,6 +22,7 @@ master: 2.5.0
 * Improved buffer checking for ``std::initializer_list``
 * Add convenience functions ``argc()`` and ``argv()`` to ``cppyy.ll``
 * Added ``nullptr`` comparisons for for typed ``nullptr``
+* Support for ``using`` pointer types as template arguments
 
 
 2022-10-03: 2.4.1